refactor theme toggling
