+++ frontend/src/hooks/useTheme.tsx
@@ -3,8 +3,20 @@ import { useEffect } from "react";
 export default function useTheme() {
   useEffect(() => {
     const stored = localStorage.getItem("theme");
-    const theme = stored ?? "light";
-
-    document.documentElement.classList.toggle("dark", theme === "dark");
+    if (!stored) {
+      localStorage.setItem("theme", "dark"); //default to dark
+      document.documentElement.classList.add("dark");
+    } else {
+      document.documentElement.classList.toggle("dark", stored === "dark");
+    }
   }, []);
+
+  const toggleTheme = () => {
+    const root = document.documentElement;
+    const isDark = root.classList.contains("dark");
+    root.classList.toggle("dark");
+    localStorage.setItem("theme", isDark ? "light" : "dark");
+  };
+
+  return { toggleTheme };
 }
